requirements.txt
Django==2.2.7
django-haystack==2.8.1
djangorestframework==3.10.3
djangorestframework-stubs==1.0.0
drf-haystack==1.8.6
elasticsearch==2.4.1
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
